Split the simulator from the API interface
Jira ticket: RSD-605
